Generalize input generation
Rather than specially process the transform-type based arguments, use
new function `generateInputs` to generate variables for every input
which is a generator/transform.

Then, in the `shaderString` function, use the variable that would be
generated by `generateInputs` as the input. An improvement would be
to make a new function which generates the variable name that can be
re-used, in order to make the dependency between these two functions
more obvious.

This renders a few different types semi-obsolete:

- `combine`: the code is exactly the same as for `color`
- `combineCoord`: the code is exactly the same as for `coord`

This makes sense, as for the purposes of chaining transforms together,
there are only 3 shapes or categories of functions:

- Functions which start the chain and produce the color var
- functions which assign to the color
- functions which assign to the uv

All inputs follow the same pattern otherwise, and it is the function
body (which is irrelevant to the shader generator) which gives those
inputs meaning.

I believe the `combine` type is still very useful for user discovery ;
however, while `combineCoord` provides similar utility, I believe it
is a less essential category than `combine`.